Adding code to make unit test in the jugar() method.

+++ src/JocTest.java
@@ -1,5 +1,5 @@
 import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Assertions;
+
 class JocTest {
 
     private Joc joc;
@@ -12,7 +12,7 @@ class JocTest {
     @org.junit.jupiter.api.Test
     void jugar() {
         // Caso de prueba 1: Jugada válida en una celda vacía
-        joc.jugar(0, "0");
+        joc.jugar((short) 0, "0");
         char[][] expectedTaulell1 = {
                 {'X', '-', '-'},
                 {'-', '-', '-'},
@@ -22,7 +22,7 @@ class JocTest {
         assertEquals(1, joc.getTorn());
 
         // Caso de prueba 2: Jugada inválida en una celda ocupada
-        joc.jugar(0, "0");
+        joc.jugar((short) 0, "0");
         char[][] expectedTaulell2 = {
                 {'X', '-', '-'},
                 {'-', '-', '-'},
@@ -32,7 +32,7 @@ class JocTest {
         assertEquals(1, joc.getTorn());
 
         // Caso de prueba 3: Jugada válida en otra celda vacía
-        joc.jugar(1, "1");
+        joc.jugar((short) 1, "1");
         char[][] expectedTaulell3 = {
                 {'X', '-', '-'},
                 {'-', 'O', '-'},
